backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0